Consolidate tests for recipes.js into one test file
