+++ test/recipes-test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { getRecipeInstructions, getRecipeById, getIngredients, getIngredientNames } from '../src/recipes.js'
+import { getRecipeInstructions, getRecipeById, filterRecipes, getIngredients, getIngredientNames } from '../src/recipes.js'
 import { sampleRecipeData } from '../src/data/sample-recipes.js';
 import { sampleIngredientsData } from '../src/data/sample-ingredients.js';
 
@@ -44,6 +44,31 @@ describe ('recipe info', () => {
     ])
   });
 });
+
+describe ('filter', function() {
+  
+  it('should return an array of filtered recipes by a tag', function() {
+    const filteredRecipes = filterRecipes(sampleRecipeData, 'starter')
+    expect(filteredRecipes).to.be.deep.equal([sampleRecipeData[0]])
+  })
+  
+  it('should be able return an array of filtered recipes by a different tag', function() {
+    const filteredRecipes = filterRecipes(sampleRecipeData, 'sauce')
+    expect(filteredRecipes).to.be.deep.equal([sampleRecipeData[2]])
+  })
+
+  it('should be able return an array of filtered recipes by a name', function() {
+    const filteredRecipes = filterRecipes(sampleRecipeData, "Dirty Steve's Original Wing Sauce")
+    expect(filteredRecipes).to.be.deep.equal([sampleRecipeData[2]]
+    )
+  })
+  
+  it('should let the user know if there were no results found', function() {
+    const filteredRecipes = filterRecipes(sampleRecipeData, 'Plastic Garbage')
+    expect(filteredRecipes).to.be.equal('Sorry, no matching results!')
+  })
+
+})
 
 describe('ingredients', () => {
   let recipe1, recipe2;
